+++ jetty/src/io/pedestal/http/jetty.clj
@@ -13,27 +13,27 @@
 
 (ns io.pedestal.http.jetty
   "Jetty adaptor for Pedestal."
-  (:require [io.pedestal.http.jetty.container]
+  (:require io.pedestal.http.jetty.container
             [clojure.string :as string]
-            [io.pedestal.internal :refer [deprecated]]
             [io.pedestal.websocket :as ws])
   (:import (jakarta.websocket.server ServerContainer)
-           (org.eclipse.jetty.server Server
+           (org.eclipse.jetty.ee10.servlet ServletContextHandler ServletHolder)
+           (org.eclipse.jetty.http2 HTTP2Cipher)
+           (org.eclipse.jetty.http2.api.server ServerSessionListener)
+           (org.eclipse.jetty.http2.server HTTP2CServerConnectionFactory RawHTTP2ServerConnectionFactory)
+           (org.eclipse.jetty.server ConnectionFactory
+                                     Server
                                      HttpConfiguration
                                      SecureRequestCustomizer
-                                     ConnectionFactory
                                      HttpConnectionFactory
-                                     ServerConnector SslConnectionFactory)
-           (org.eclipse.jetty.servlet ServletContextHandler ServletHolder)
+                                     ServerConnector
+                                     SslConnectionFactory)
            (org.eclipse.jetty.util.thread QueuedThreadPool ThreadPool)
            (org.eclipse.jetty.util.ssl SslContextFactory SslContextFactory$Server)
            (org.eclipse.jetty.alpn.server ALPNServerConnectionFactory)
-           (org.eclipse.jetty.http2 HTTP2Cipher)
-           (org.eclipse.jetty.http2.server HTTP2ServerConnectionFactory
-                                           HTTP2CServerConnectionFactory)
            (jakarta.servlet Servlet ServletContext)
            (java.security KeyStore)
-           (org.eclipse.jetty.websocket.jakarta.server.config JakartaWebSocketServletContainerInitializer JakartaWebSocketServletContainerInitializer$Configurator)))
+           (org.eclipse.jetty.ee10.websocket.jakarta.server.config JakartaWebSocketServletContainerInitializer JakartaWebSocketServletContainerInitializer$Configurator)))
 
 ;; Implement any container specific optimizations from Pedestal's container protocols
 
@@ -81,20 +81,22 @@
 
 (defn- http-configuration
   "Provides an HttpConfiguration that can be consumed by connection factories.
-  The `:io.pedestal.http.jetty/http-configuration` option can be used to specify
+  The `:http-configuration` option can be used to specify
   your own HttpConfiguration instance."
   ^HttpConfiguration [options]
-  (if-let [http-conf-override ^HttpConfiguration (::http-configuration options)]
-    http-conf-override
-    (let [{:keys [insecure-ssl?]} options
-          http-conf ^HttpConfiguration (HttpConfiguration.)]
-      (doto http-conf
-        (.setSendDateHeader true)
-        (.setSendXPoweredBy false)
-        (.setSendServerVersion false)
-        ;; :insecure-ssl? is useful for local development, as otherwise "localhost"
-        ;; is not allowed as a valid host name, resulting in 400 statuses.
-        (.addCustomizer (SecureRequestCustomizer. (not insecure-ssl?)))))))
+  (or (:http-configuration options)
+      ;; In 0.7 and earlier, was a namespaced key
+      (::http-configuration options)
+      ;; Otherwise, build on the fly
+      (let [{:keys [insecure-ssl?]} options
+            http-conf ^HttpConfiguration (HttpConfiguration.)]
+        (doto http-conf
+          (.setSendDateHeader true)
+          (.setSendXPoweredBy false)
+          (.setSendServerVersion false)
+          ;; :insecure-ssl? is useful for local development, as otherwise "localhost"
+          ;; is not allowed as a valid host name, resulting in 400 statuses.
+          (.addCustomizer (SecureRequestCustomizer. (not insecure-ssl?)))))))
 
 (defn- needed-pool-size
   "Jetty calculates a needed number of threads per acceptors and selectors,
@@ -103,7 +105,6 @@
   number of threads are created for the server."
   ([]
    (let [cores     (.availableProcessors (Runtime/getRuntime))
-         ;; TODO: Is this still accurate for Jetty 11?
          ;; The Jetty docs claim acceptors is 1.5 the number of cores available,
          ;; but the code says:  1 + cores / 16 - https://github.com/eclipse/jetty.project/blob/master/jetty-server/src/main/java/org/eclipse/jetty/server/AbstractConnector.java#L192
          acceptors (int (* 1.5 cores))                      ;(inc (/ cores 16))
@@ -123,60 +124,79 @@
   (or thread-pool
       (QueuedThreadPool. ^Integer max-threads)))
 
+(defn- add-connection-factories
+  [^Server server factories]
+  (let [factories' (into-array ConnectionFactory (remove nil? factories))
+        conn       (ServerConnector. server factories')]
+    (.addConnector server conn)
+    ;; Return the ServerConnector for any further configuration
+    conn))
+
 ;; Consider allowing users to set the number of acceptors (ideal at 1 per core) and/or selectors
 (defn- create-server
   "Construct a Jetty Server instance."
   [servlet options]
   (let [{:keys [host port websockets container-options]} options
-        {:keys [ssl? ssl-port
+        {:keys [ssl? ssl-port max-streams
                 h2? h2c? connection-factory-fns
                 context-configurator context-path configurator daemon? reuse-addr?]
          :or   {configurator identity
                 context-path "/"
                 h2c?         true
+                max-streams  128
                 reuse-addr?  true}} container-options
         ^ThreadPool thread-pool (thread-pool options)
         server                  (Server. thread-pool)
-        _                       (when-not (string/starts-with? context-path "/")
-                                  (throw (IllegalArgumentException. "context-path must begin with a '/'")))
-        _                       (when (and h2? (not ssl-port))
-                                  (throw (IllegalArgumentException. "SSL must be enabled to use HTTP/2. Please set an ssl port and appropriate *store setups")))
-        _                       (when (and (nil? port) (not (or ssl? ssl-port h2?)))
-                                  (throw (IllegalArgumentException. "HTTP was turned off with a `nil` port value, but no SSL config was supplied.  Please set an HTTP port or configure SSL")))
-        _                       (when (and (nil? port) (true? h2c?))
-                                  (throw (IllegalArgumentException. "HTTP was turned off with a `nil` port value, but you attempted to turn on HTTP2-Cleartext.  Please set an HTTP port or set `h2c?` to false in your service config")))
+        _                       (do
+                                  (when-not (string/starts-with? context-path "/")
+                                    (throw (ex-info "context-path must begin with a '/'" {:container-options container-options})))
+
+                                  (when (and h2? (not ssl-port))
+                                    (throw (ex-info "SSL must be enabled to use HTTP/2; Provide keys :ssl-port and keystore/truststore configuration"
+                                                    {:container-options container-options})))
+
+                                  (when (and (nil? port) (not (or ssl? ssl-port h2?)))
+                                    (throw (ex-info "No HTTP or SSL port configured"
+                                                    {:container-options container-options})))
+
+                                  (when (and (nil? port)
+                                             h2c?)
+                                    (throw (ex-info "HTTP2-Cleartext can not be enabled unless a non-nil HTTP port is provided"
+                                                    {:container-options container-options}))))
+        server-session-listener (reify ServerSessionListener)
         http-conf               (http-configuration container-options)
         http                    (HttpConnectionFactory. http-conf)
-        http2c                  (when h2c? (HTTP2CServerConnectionFactory. http-conf))
-        http2                   (when h2? (HTTP2ServerConnectionFactory. http-conf))
+        http2c                  (when h2c?
+                                  (doto (HTTP2CServerConnectionFactory. http-conf)
+                                    (.setMaxConcurrentStreams max-streams)))
+        http2                   (when h2?
+                                  (doto (RawHTTP2ServerConnectionFactory. server-session-listener)
+                                    (.setMaxConcurrentStreams max-streams)
+                                    (.setConnectProtocolEnabled true)))
         alpn                    (when h2?
-                                  ;(NegotiatingServerConnectionFactory/checkProtocolNegotiationAvailable) ;; This only looks at Java8 bootclasspath stuff, and is no longer valid in newer Jetty versions
+                                  ;; Application-Layer Protocol Negotiation
                                   (doto (ALPNServerConnectionFactory. "h2,h2-17,h2-14,http/1.1")
                                     (.setDefaultProtocol "http/1.1")))
         ssl                     (when (or ssl? ssl-port h2?)
                                   (ssl-conn-factory (assoc options :alpn alpn)))
         http-connector          (when port
-                                  (doto (ServerConnector. server (into-array ConnectionFactory
-
-                                                                             (remove nil? [http http2c])))
+                                  (doto (add-connection-factories server [http http2c])
                                     (.setReuseAddress reuse-addr?)
                                     (.setPort port)
                                     (.setHost host)))
         ssl-connector           (when ssl
-                                  (doto (ServerConnector. server
-                                                          (into-array ConnectionFactory
-                                                                      (remove nil?
-                                                                              (into [ssl alpn http2 (HttpConnectionFactory. http-conf)]
-                                                                                    (map (fn [ffn] (ffn options http-conf)) connection-factory-fns)))))
-                                    (.setReuseAddress reuse-addr?)
-                                    (.setPort ssl-port)
-                                    (.setHost host)))
-        servlet-context-handler (doto (ServletContextHandler. server context-path)
+                                  (let [factories (into [ssl alpn http2 (HttpConnectionFactory. http-conf)]
+                                                        (map (fn [ffn] (ffn options http-conf)) connection-factory-fns))]
+                                    (doto (add-connection-factories server factories)
+                                      (.setReuseAddress reuse-addr?)
+                                      (.setPort ssl-port)
+                                      (.setHost host))))
+        servlet-context-handler (doto (ServletContextHandler. ServletContextHandler/SESSIONS)
+                                  (.setContextPath context-path)
                                   (.addServlet (ServletHolder. ^Servlet servlet) "/*"))]
     (when websockets
       (JakartaWebSocketServletContainerInitializer/configure servlet-context-handler
                                                              (reify JakartaWebSocketServletContainerInitializer$Configurator
-
                                                                (^void accept [_this ^ServletContext _context
                                                                               ^ServerContainer container]
                                                                  (ws/add-endpoints container websockets)))))
@@ -190,69 +210,30 @@
       (.addConnector server ssl-connector))
     (when context-configurator
       (context-configurator servlet-context-handler))
+
+    ;; Only set the handler once it is fully configured
+    (.setDefaultHandler server servlet-context-handler)
+
+    ;; And only after that perform final configuration of the server
     (configurator server)))
 
 
-(defn- -start
-  "Deprecated; to be made private in the future."
+(defn- start
   [^Server server
    {:keys [join?] :or {join? true}}]
   (.start server)
   (when join? (.join server))
   server)
 
-(defn ^{:deprecated "0.7.0"} start
-  "Deprecated; to be made private in the future."
-  [^Server server options]
-  (deprecated `start
-    (-start server options)))
-
-(defn- -stop [^Server server]
+(defn- stop
+  [^Server server]
   (.stop server)
   server)
 
-(defn ^{:deprecated "0.7.0"} stop
-  "Deprecated; to be made private in the future."
-  [^Server server]
-  (deprecated `stop
-    (-stop server)))
-
 (defn server
   "Called from [[io.pedestal.http/server]] to create a Jetty server instance."
-  ([service-map] (server service-map {}))
-  ([service-map options]
-   (let [server (create-server (:io.pedestal.http/servlet service-map) options)]
-     {:server   server
-      :start-fn #(-start server options)
-      :stop-fn  #(-stop server)})))
-
-
-;; TODO: spec all this
-
-;; :port         - the http/h2c port to listen on (defaults to 80);
-;;                     If nil and SSL Config is set, HTTP is disabled.
-;; :host         - the hostname to listen on
-;; :join?        - blocks the thread until server ends (defaults to true)
-
-;; -- Container Options --
-;; :daemon?      - use daemon threads (defaults to false)
-;; :max-threads  - the maximum number of threads to use (default 50)
-;; :thread-pool  - override the Jetty thread pool (ignores max-threads)
-;; :reuse-addr?  - reuse the socket address (defaults to true)
-;; :configurator - a function called with the Jetty Server instance
-;; :context-configurator - a function called with the Jetty ServletContextHandler
-;; :ssl?         - allow connections over HTTPS
-;; :insecure-ssl? - if true, then SNI check is disabled (useful for http locally with host "localhost")
-;; :ssl-port     - the SSL port to listen on (defaults to 443, implies :ssl?)
-;; :h2?          - enable http2 protocol on secure socket port
-;; :h2c?         - enable http2 clear text on plain socket port
-;; :connection-factory-fns - a vector of functions that take the options map and HttpConfiguration
-;;                           and return a ConnectionFactory obj (applied to SSL connection)
-;; :ssl-context-factory - a Jetty SslContextFactory to use in place of one constructed by Pedestal.
-;;                        If passed in, all other SSL related options are ignored.
-;; :keystore     - the keystore to use for SSL connections - may be the object or a string path
-;; :key-password - the password to the keystore
-;; :truststore   - a truststore to use for SSL connections - may be the object or a string path
-;; :trust-password - the password to the truststore
-;; :client-auth  - SSL client certificate authenticate, may be set to :need,
-;;                 :want or :none (defaults to :none)"
+  [service-map options]
+  (let [server (create-server (:io.pedestal.http/servlet service-map) options)]
+    {::server   server
+     :start-fn #(start server options)
+     :stop-fn  #(stop server)}))